moved isFrameExtension into origins module
